+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### Bug fixes
 
-- Only validate a path if it is not null ([#50](https://github.com/nextflow-io/nf-validation/pull/50/files))
-- Only validate a file with a schema if the file path is provided ([#51](https://github.com/nextflow-io/nf-validation/pull/51/files))
+- Only validate a path if it is not null ([#50](https://github.com/nextflow-io/nf-validation/pull/50))
+- Only validate a file with a schema if the file path is provided ([#51](https://github.com/nextflow-io/nf-validation/pull/51))
+- Handle errors when sample sheet not provided or doesn't have a schema ([#56](https://github.com/nextflow-io/nf-validation/pull/56))
 
 ## Version 0.2.1