+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Correctly handle double-quoted fields containing commas in csv files by `.fromSamplesheet()` ([#63](https://github.com/nextflow-io/nf-validation/pull/63))
 - Print param name when path does not exist ([#65](https://github.com/nextflow-io/nf-validation/pull/65))
 - Fix file or directory does not exist error not printed when it was the only error in a samplesheet ([#65](https://github.com/nextflow-io/nf-validation/pull/65))
+- Do not return parameter in summary if it has no default in the schema and is set to 'false' ([#66](https://github.com/nextflow-io/nf-validation/pull/66))
 
 ## Version 0.2.1