+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # nextflow-io/nf-validation: Changelog
 
-## Version 0.3.0 (dev)
+## Version 0.3.0
 
 ### Bug fixes